requirements.txt
matplotlib==3.7.2
num2words==0.5.13
numpy==1.24.1
pandas==2.0.3
Pillow==9.3.0
Pillow==10.1.0
pycocotools==2.0.7
Requests==2.31.0
salesforce_lavis==1.0.1
scikit_learn==1.3.0
scipy==1.11.3
thefuzz==0.20.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.31.0
wandb==0.15.8
salesforce-lavis